Change name to customer_pricing_in_product_view and exclude action_open from test Updates as per request from Pedro Baeza
