--- customer_context_in_product_view/__manifest__.py
@@ -1,33 +1,0 @@
-# -*- coding: utf-8 -*-
-##############################################################################
-#
-#    OpenERP, Open Source Management Solution - module extension
-#    Copyright (C) 2014- O4SB (<http://o4sb.com>).
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
-
-{
-    'name': 'Display Customer Price in Product View',
-    'version': '10.0.1.0.0',
-    'category': 'Sales',
-    'author': "Odoo Community Association (OCA)",
-    'website': 'http://odoo-community.org',
-    'license': 'AGPL-3',
-    'depends': ['product', 'sales_team'],
-    'data': ['partner_pricelist_view.xml'],
-    'installable': True,
-    'active': False,
-}
