+++ customer_pricing_in_product_view/__manifest__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+# Copyright 2014 Odoo Community Association (OCA), O4SB - Graeme Gellatly
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+{
+    'name': 'Display Customer Price in Product View',
+    'version': '10.0.1.0.0',
+    'category': 'Sales',
+    'author': "Odoo Community Association (OCA), O4SB - Graeme Gellatly",
+    'website': 'http://odoo-community.org',
+    'license': 'AGPL-3',
+    'depends': ['product', 'sales_team'],
+    'data': ['views/res_partner_view.xml'],
+    'installable': True,
+    'active': False,
+}
