+++ customer_pricing_in_product_view/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+# Copyright 2014 Odoo Community Association (OCA), O4SB - Graeme Gellatly
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
+from . import res_partner
