+++ customer_pricing_in_product_view/views/res_partner_view.xml
